# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Rent-a-Car API",
    "version": "1.0.0",
    "description": "API endpoints for user authentication, car browsing, and rental slot booking."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "paths": {
    "/api/auth/register": {
      "post": {
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "example": "vuk@example.com" },
                  "password": { "type": "string", "example": "password123" },
                  "name": { "type": "string", "example": "Vuk Ivanovic" }
                },
                "required": ["email", "password", "name"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User successfully registered",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "user": {
                      "type": "object",
                      "properties": {
                        "email": { "type": "string" },
                        "name": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "User already exists or missing parameters" }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Login existing user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "example": "vuk@example.com" },
                  "password": { "type": "string", "example": "password123" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "token": { "type": "string" },
                    "user": {
                      "type": "object",
                      "properties": {
                        "email": { "type": "string" },
                        "name": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/api/cars": {
      "get": {
        "summary": "Retrieve available cars",
        "responses": {
          "200": {
            "description": "A list of cars",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "string" },
                      "name": { "type": "string" },
                      "type": { "type": "string" },
                      "transmission": { "type": "string" },
                      "power": { "type": "string" },
                      "acceleration": { "type": "string" },
                      "price": { "type": "integer" },
                      "description": { "type": "string" },
                      "features": { "type": "array", "items": { "type": "string" } }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/bookings": {
      "get": {
        "summary": "Get all bookings (or filter by logged-in user email)",
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "required": true,
            "schema": { "type": "string" },
            "description": "Email of user to filter bookings"
          }
        ],
        "responses": {
          "200": {
            "description": "List of user bookings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "string" },
                      "carId": { "type": "string" },
                      "carName": { "type": "string" },
                      "startDate": { "type": "string" },
                      "endDate": { "type": "string" },
                      "totalPrice": { "type": "number" }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new booking slot",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "carId": { "type": "string", "example": "tesla-s" },
                  "userEmail": { "type": "string", "example": "vuk@example.com" },
                  "startDate": { "type": "string", "example": "2026-07-10" },
                  "endDate": { "type": "string", "example": "2026-07-15" }
                },
                "required": ["carId", "userEmail", "startDate", "endDate"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Booking successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "booking": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string" },
                        "carId": { "type": "string" },
                        "startDate": { "type": "string" },
                        "endDate": { "type": "string" },
                        "totalPrice": { "type": "number" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Dates already taken, invalid dates format, or missing info" }
        }
      }
    }
  }
}